Trait fiat_filter::Config

pub trait Config: Config + Trait + Trait + Trait + Trait + Trait {
    type PriceProvider: PriceProvider;
    type Currency: Currency<Self::AccountId>;
    type Call: Parameter + UnfilteredDispatchable<Origin = Self::Origin> + GetDispatchInfo + IsSubType<Call<Self>> + IsSubType<Call<Self>> + IsSubType<Call<Self>> + IsSubType<Call<Self>> + IsSubType<Call<Self>>;
    type MinDockFiatRate: Get<u32>;
}

The pallet’s configuration trait Configure the pallet by specifying the parameters and types on which it depends.

Associated Types

type PriceProvider: PriceProvider

Config option for updating the DockFiatRate

type Currency: Currency<Self::AccountId>

The module’s Currency type definition

type Call: Parameter + UnfilteredDispatchable<Origin = Self::Origin> + GetDispatchInfo + IsSubType<Call<Self>> + IsSubType<Call<Self>> + IsSubType<Call<Self>> + IsSubType<Call<Self>> + IsSubType<Call<Self>>

The outer call (dispatchable) that this module is called with. It is possible to use another type here, but it’s expected that your runtime::Call will be used. Since we need to match by origin crate, we implement IsSubType for each Call type we want to match for.

type MinDockFiatRate: Get<u32>

Minimum price, in case the price fetched from optimized_get_dock_usd_price is zero or an error expressed in USD_1000th/DOCK (as u32) (== USD/1000DOCK) just like the actual result from optimized_get_dock_usd_price